fix: support decimal meter readings for accurate utility billing
- Changed MeterReading entity to use double instead of int for opening/closing readings
- Updated input formatters to allow decimal values (up to 2 decimal places)
- Modified sliding scale calculations to handle decimal units
- Updated TariffStep to use double for upToUnits
- Fixed all calculation logic in new bill screen, main list, and PDF export
- Updated test files to use double values
- Added proper validation for decimal inputs

This fixes the critical issue where decimal meter readings (e.g., 2.4kl to 2.6kl = 0.2kl)
were being truncated to integers, resulting in lost billing units.

+++ lib/screens/main_bill_list_screen.dart
@@ -201,14 +201,14 @@ class MainBillListScreen extends ConsumerWidget {
                         
                         // Water sliding scale calculation
                         double waterCost = 0;
-                        int remainingUnits = waterUnits;
+                        double remainingUnits = waterUnits;
                         if (remainingUnits > 0) {
-                          final int firstTier = remainingUnits > 6 ? 6 : remainingUnits;
+                          final double firstTier = remainingUnits > 6 ? 6 : remainingUnits;
                           waterCost += firstTier * waterTariff0to6;
                           remainingUnits -= firstTier;
                         }
                         if (remainingUnits > 0) {
-                          final int secondTier = remainingUnits > 9 ? 9 : remainingUnits;
+                          final double secondTier = remainingUnits > 9 ? 9 : remainingUnits;
                           waterCost += secondTier * waterTariff7to15;
                           remainingUnits -= secondTier;
                         }
@@ -220,12 +220,12 @@ class MainBillListScreen extends ConsumerWidget {
                         double sanitationCost = 0;
                         remainingUnits = sanitationUnits;
                         if (remainingUnits > 0) {
-                          final int firstTier = remainingUnits > 6 ? 6 : remainingUnits;
+                          final double firstTier = remainingUnits > 6 ? 6 : remainingUnits;
                           sanitationCost += firstTier * sanitationTariff0to6;
                           remainingUnits -= firstTier;
                         }
                         if (remainingUnits > 0) {
-                          final int secondTier = remainingUnits > 9 ? 9 : remainingUnits;
+                          final double secondTier = remainingUnits > 9 ? 9 : remainingUnits;
                           sanitationCost += secondTier * sanitationTariff7to15;
                           remainingUnits -= secondTier;
                         }
